CancellationTokenSource -> CancellationToken
There's no need for APIs to receive the CancellationTokenSource
object unless the library wants to take the decision of cancel
the work by itself.

But if who takes the decision is always the consumer of the library
then only CancellationToken is needed (and its default value doesn't
need to be null because CancellationToken.None exists).